dev-5931 migrate helpers and data points out of sticky header to their respective files
