+++ src/js/components/about/legal/common/LegalContent.jsx
@@ -7,7 +7,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import Sidebar from 'components/sharedComponents/sidebar/Sidebar';
-import { stickyHeaderHeight } from 'components/sharedComponents/stickyHeader/StickyHeader';
+import { stickyHeaderHeight } from 'dataMapping/stickyHeader/stickyHeader';
 
 const legalSections = [
     {
