+++ src/js/components/covid19/DataSourcesAndMethodologiesPage.jsx
@@ -22,7 +22,9 @@ import Footer from 'containers/Footer';
 import Header from 'containers/shared/HeaderContainer';
 import { useDefCodes } from 'containers/covid19/WithDefCodes';
 import MetaTags from 'components/sharedComponents/metaTags/MetaTags';
-import StickyHeader, { useDynamicStickyClass, getStickyBreakPointForSidebar, stickyHeaderHeight } from 'components/sharedComponents/stickyHeader/StickyHeader';
+import StickyHeader from 'components/sharedComponents/stickyHeader/StickyHeader';
+import { stickyHeaderHeight } from 'dataMapping/stickyHeader/stickyHeader';
+import { getStickyBreakPointForSidebar, useDynamicStickyClass } from 'helpers/stickyHeaderHelper';
 import Sidebar from 'components/sharedComponents/sidebar/Sidebar';
 import ExternalLink from 'components/sharedComponents/ExternalLink';
 import ShareIcon from 'components/sharedComponents/stickyHeader/ShareIcon';
