+++ src/js/components/sharedComponents/stickyHeader/StickyHeader.jsx
@@ -3,65 +3,13 @@
  * Created by Mike Bray 02/02/2018
  **/
 
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useRef, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { throttle } from 'lodash';
-import Cookies from 'js-cookie';
+import { useDynamicStickyClass } from 'helpers/stickyHeaderHelper';
 
 const propTypes = {
     children: PropTypes.node
 };
-
-export const stickyHeaderHeight = 66;
-export const globalBannerHeight = 90;
-export const globalCovidBannerCookie = 'usaspending_covid_release';
-
-export const getStickyBreakPointForSidebar = () => {
-    const isGlobalBannerHidden = Cookies.get(globalCovidBannerCookie) === 'hide';
-
-    if (isGlobalBannerHidden) {
-        return 97;
-    }
-    return 97 + globalBannerHeight;
-};
-
-export const useDynamicStickyClass = (stickyRef, fixedStickyBreakpoint = null) => {
-    const [dynamicStickyBreakpoint, setDynamicStickyBreakpoint] = useState(0);
-    const [isSticky, setIsSticky] = useState(false);
-    return [
-        isSticky,
-        // scrollPosition at which we apply the sticky-class
-        dynamicStickyBreakpoint,
-        // setSticky
-        setIsSticky,
-        // handleScroll
-        throttle(() => {
-            const scrollY = window.scrollY || document.documentElement.scrollTop;
-            if (fixedStickyBreakpoint && scrollY >= fixedStickyBreakpoint && !isSticky) {
-                // we know which y position to apply the sticky class
-                setIsSticky(true);
-            }
-            else if (!fixedStickyBreakpoint && scrollY >= dynamicStickyBreakpoint && !isSticky) {
-                // we don't know which y position to apply the sticky class
-                setIsSticky(true);
-            }
-            else if (scrollY < fixedStickyBreakpoint && isSticky) {
-                setIsSticky(false);
-            }
-            else if (scrollY < dynamicStickyBreakpoint && isSticky) {
-                setIsSticky(false);
-            }
-        }, 100),
-        // measureScreen
-        throttle(() => {
-            const wrapperY = stickyRef.current
-                ? stickyRef.current.offsetTop
-                : 0;
-            setDynamicStickyBreakpoint(wrapperY);
-        }, 100)
-    ];
-};
-
 
 const StickyHeader = ({
     children
