+++ src/js/containers/covid19/Covid19Container.jsx
@@ -12,7 +12,9 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import MetaTags from 'components/sharedComponents/metaTags/MetaTags';
 import Header from 'containers/shared/HeaderContainer';
 import Sidebar from 'components/sharedComponents/sidebar/Sidebar';
-import StickyHeader, { useDynamicStickyClass, getStickyBreakPointForSidebar, stickyHeaderHeight } from 'components/sharedComponents/stickyHeader/StickyHeader';
+import StickyHeader from 'components/sharedComponents/stickyHeader/StickyHeader';
+import { stickyHeaderHeight } from 'dataMapping/stickyHeader/stickyHeader';
+import { getStickyBreakPointForSidebar, useDynamicStickyClass } from 'helpers/stickyHeaderHelper';
 import Covid19Section from 'components/covid19/Covid19Section';
 import Footer from 'containers/Footer';
 import Heading from 'components/covid19/Heading';
