+++ src/js/dataMapping/stickyHeader/stickyHeader.js
@@ -1,0 +1,3 @@
+export const stickyHeaderHeight = 66;
+export const globalBannerHeight = 90;
+export const globalCovidBannerCookie = 'usaspending_covid_release';
